Refactor SQL implementations
