+++ lumen/sources/snowflake.py
@@ -1,16 +1,10 @@
 from __future__ import annotations
 
-from typing import Any
-
-import pandas as pd
 import param
 import snowflake.connector
 
-from ..transforms.sql import (
-    SQLDistinct, SQLFilter, SQLLimit, SQLMinMax,
-)
-from ..util import get_dataframe_schema
-from .base import BaseSQLSource, cached, cached_schema
+from ..transforms.sql import SQLFilter
+from .base import BaseSQLSource, cached
 
 
 class SnowflakeSource(BaseSQLSource):
@@ -88,10 +82,13 @@ class SnowflakeSource(BaseSQLSource):
         params['conn'] = self._conn
         return SnowflakeSource(**params)
 
+    def execute(self, sql_query: str):
+        return self._cursor.execute(sql_query).fetch_pandas_all()
+
     def get_tables(self) -> list[str]:
         if isinstance(self.tables, dict | list):
             return list(self.tables)
-        tables = self._cursor.execute(f'SELECT TABLE_NAME, TABLE_SCHEMA FROM {self.database}.INFORMATION_SCHEMA.TABLES;').fetch_pandas_all()
+        tables = self.execute(f'SELECT TABLE_NAME, TABLE_SCHEMA FROM {self.database}.INFORMATION_SCHEMA.TABLES;')
         return [f'{self.database}.{row.TABLE_SCHEMA}.{row.TABLE_NAME}' for _, row in tables.iterrows()]
 
     def get_sql_expr(self, table: str):
@@ -113,57 +110,4 @@ class SnowflakeSource(BaseSQLSource):
             sql_transforms = [SQLFilter(conditions=conditions)] + sql_transforms
         for st in sql_transforms:
             sql_expr = st.apply(sql_expr)
-        return self._cursor.execute(sql_expr).fetch_pandas_all()
-
-    @cached_schema
-    def get_schema(
-        self, table: str | None = None, limit: int | None = None
-    ) -> dict[str, dict[str, Any]] | dict[str, Any]:
-        if table is None:
-            tables = self.get_tables()
-        else:
-            tables = [table]
-
-        schemas = {}
-        sql_limit = SQLLimit(limit=limit or 1)
-        for entry in tables:
-            sql_expr = self.get_sql_expr(entry)
-            data = self._cursor.execute(sql_limit.apply(sql_expr)).fetch_pandas_all()
-            schemas[entry] = schema = get_dataframe_schema(data)['items']['properties']
-            if limit:
-                continue
-
-            enums, min_maxes = [], []
-            for name, col_schema in schema.items():
-                if 'enum' in col_schema:
-                    enums.append(name)
-                elif 'inclusiveMinimum' in col_schema:
-                    min_maxes.append(name)
-            for col in enums:
-                distinct_expr = SQLDistinct(columns=[col]).apply(sql_expr)
-                distinct_expr = ' '.join(distinct_expr.splitlines())
-                distinct = self._cursor.execute(distinct_expr).fetch_pandas_all()
-                schema[col]['enum'] = distinct[col].tolist()
-
-            if not min_maxes:
-                continue
-
-            minmax_expr = SQLMinMax(columns=min_maxes).apply(sql_expr)
-            minmax_expr = ' '.join(minmax_expr.splitlines())
-            minmax_data = self._cursor.execute(minmax_expr).fetch_pandas_all()
-            for col in min_maxes:
-                kind = data[col].dtype.kind
-                if kind in 'iu':
-                    cast = int
-                elif kind == 'f':
-                    cast = float
-                elif kind == 'M':
-                    cast = str
-                else:
-                    cast = lambda v: v
-                min_data = minmax_data[f'{col}_min'].iloc[0]
-                schema[col]['inclusiveMinimum'] = min_data if pd.isna(min_data) else cast(min_data)
-                max_data = minmax_data[f'{col}_max'].iloc[0]
-                schema[col]['inclusiveMaximum'] = max_data if pd.isna(max_data) else cast(max_data)
-
-        return schemas if table is None else schemas[table]
+        return self.execute(sql_expr)
